% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coretta2017egg.R
\docType{package}
\name{coretta2017egg}
\alias{coretta2017egg}
\alias{coretta2017egg-package}
\title{coretta2017egg: Data package of a pilot study of EGG data analysis}
\description{
Provides two datasets of a pilot study of EGG data analysis.
}
