% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracegram_p}
\alias{tracegram_p}
\title{Pilot tracegram data}
\format{A tibble with 5999 observations and 7 variables.
\describe{
 \item{\code{speaker}}{speaker ID}
 \item{\code{ponation}}{phonation (\code{modal}, \code{breathy})}
 \item{\code{token}}{numeric ID of the token (utterance)}
 \item{\code{time}}{proportion of the duration of the utterance}
 \item{\code{egg_minimum}}{normalised time (proportion) of EGG minimum}
 \item{\code{degg_maximum}}{normalised time (proportion) of the dEGG maximum}
 \item{\code{degg_minimum}}{normalised time (proportion) of the dEGG minimum}
}}
\usage{
tracegram_p
}
\description{
A dataset with tracegram data from 5 speakers from a pilot study.
}
\keyword{datasets}
