% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wavegram_p}
\alias{wavegram_p}
\title{Pilot wavegram data}
\format{A tibble with 31864 observations of 7 variables.
\describe{
 \item{\code{speaker}}{speaker ID}
 \item{\code{ponation}}{phonation (\code{modal}, \code{breathy})}
 \item{\code{token}}{numeric ID of the token (utterance)}
 \item{\code{time}}{proportion of the duration of the utterance}
 \item{\code{sequence}}{index of glottal cycle within utterance}
 \item{\code{sample}}{sample within glottal cycle}
 \item{\code{amplitude}}{relative amplitude of the electroglottographic signal}
}}
\usage{
wavegram_p
}
\description{
A dataset with wavegram data from 5 speakers from a pilot study.
}
\keyword{datasets}
